Support fractional gauge scales and correct bar pressure dial

+++ include/gauge.h
@@ -34,13 +34,13 @@ class Gauge : public UiElement {
         float m_thetaMin;
         float m_thetaMax;
 
-        int m_min;
-        int m_max;
+        float m_min;
+        float m_max;
         int m_maxMinorTick;
         float m_gamma;
 
-        int m_minorStep;
-        int m_majorStep;
+        float m_minorStep;
+        float m_majorStep;
 
         float m_minorTickWidth;
         float m_majorTickWidth;
